Extract callback authentication to strategy class
