+++ PubSub.Application/Program.cs
@@ -40,6 +40,7 @@ builder.Services.AddSingleton<IMessageSerializer, UTF8MessageSerializer>();
 builder.Services.AddSingleton<ISubscriberNotifierService, HttpSubscriberNotifierService>();
 builder.Services.AddSingleton<ISubscriptionStore, InMemorySubscriptionStore>();
 builder.Services.AddSingleton<IRabbitMQConsumerFactory, RabbitMQConsumerFactory>();
+builder.Services.AddSingleton<ICallbackAuthenticator, HmacCallbackAuthenticator>();
 builder.Services.AddRequestMapping();
