+++ PubSub.Core/Services/CallbackAuthenticator/ICallbackAuthenticator.cs
@@ -1,0 +1,7 @@
+﻿namespace PubSub.Core.Services.CallbackAuthentication
+{
+    public interface ICallbackAuthenticator
+    {
+        string GetAuthentication(string source);
+    }
+}
